requirements.txt
mcp>=1.0.0
llama-cpp-python>=0.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
